Fix issue caused by recent commit (MTTF)

+++ sources/ckps.c
@@ -39,7 +39,7 @@
 #include "magnitude.h"
 #include "tables.h"     //fnptr_t
 
-#include "knock.h"
+#include "knock.h"                            
 
 #ifdef STROBOSCOPE
 #define STROBE_PW 31    //!< Strobe pulse width (100uS), value in tics ot timer, 1 tick = 3.2uS
@@ -63,7 +63,7 @@
  *      for 36-1 crank wheel, p * 1.5
  *      for 12-3 crank wheel, p * 3.0
  */
-#define CKPS_GAP_BARRIER(p)  (((p) * ckps.mttf) >> 8)
+#define CKPS_GAP_BARRIER(p)  (((uint32_t)(p) * ckps.mttf) >> 8)
 
 //Define values for controlling of outputs
 #define IGN_OUTPUTS_INIT_VAL 1        //!< value used for initialization
